refactor: splits out the tests into dedicated files
