+++ packages/slackBotFunction/tests/conftest.py
@@ -1,0 +1,51 @@
+import pytest
+
+from unittest.mock import Mock, patch
+from moto import mock_aws
+import boto3
+import os
+
+
+TEST_BOT_TOKEN = "test-bot-token"
+TEST_SIGNING_SECRET = "test-signing-secret"
+
+
+@pytest.fixture
+def mock_env():
+    """Mock environment variables"""
+    env_vars = {
+        "SLACK_BOT_TOKEN_PARAMETER": "/test/bot-token",
+        "SLACK_SIGNING_SECRET_PARAMETER": "/test/signing-secret",
+        "SLACK_BOT_STATE_TABLE": "test-bot-state-table",
+        "KNOWLEDGEBASE_ID": "test-kb-id",
+        "RAG_MODEL_ID": "test-model-id",
+        "AWS_REGION": "eu-west-2",
+        "GUARD_RAIL_ID": "test-guard-id",
+        "GUARD_RAIL_VERSION": "1",
+        "AWS_LAMBDA_FUNCTION_NAME": "test-function",
+    }
+    with patch.dict(os.environ, env_vars):
+        yield env_vars
+
+
+@pytest.fixture
+def mock_dynamodb_table():
+    """Mock DynamoDB table"""
+    with mock_aws():
+        dynamodb = boto3.resource("dynamodb", region_name="eu-west-2")
+        table = dynamodb.create_table(
+            TableName="test-bot-state-table",
+            KeySchema=[{"AttributeName": "eventId", "KeyType": "HASH"}],
+            AttributeDefinitions=[{"AttributeName": "eventId", "AttributeType": "S"}],
+            BillingMode="PAY_PER_REQUEST",
+        )
+        yield table
+
+
+@pytest.fixture
+def lambda_context():
+    """Mock Lambda context"""
+    context = Mock()
+    context.function_name = "test-function"
+    context.aws_request_id = "test-request-id"
+    return context
